requirements.txt
# Core LLM & RAG Frameworks
streamlit>=1.35.0,<2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Google GenAI
google-genai>=1.0.0,<2.0.0

# LangChain Ecosystem (0.3.x 스타일 - 최대 호환성)
langchain>=0.2.0,<1.0.0
langchain-community>=0.2.0,<1.0.0
langchain-core>=0.3.27,<0.4.0
langchain-text-splitters>=0.3.0,<1.0.0
langchain-google-genai>=1.0.0,<5.0.0
langchain-huggingface>=0.1.0,<1.0.0
langchain-ollama>=0.2.0,<1.0.0

# Ollama Client
ollama>=0.4.4,<1.0.0

# Vector Stores & Retrieval
faiss-cpu>=1.7.4,<2.0.0
rank_bm25>=0.2.2,<1.0.0

# Embeddings & Transformers
sentence-transformers>=2.2.2,<4.0.0
transformers>=4.30.0,<5.0.0
torch==2.5.1

# Dependency Lock (충돌 방지)
fsspec>=2023.1.0,<=2024.12.0
rich>=13.3.5,<14.0.0

# Multi-User Server (FastAPI) — 파일 관리 서버 및 클라이언트
# 단일 파일 최대 100MB / 워크스페이스 전체 최대 100MB / 비활성 24h 자동 삭제
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
httpx>=0.27.0,<1.0.0

# Claude Agent SDK (Python 3.10+, 선택 설치 — 미설치 시 Ollama/Gemini만 사용)
claude-agent-sdk>=0.1.50,<0.2.0
mcp>=1.0.0,<2.0.0

# LLM Proxy (Claude API → Ollama/OpenRouter/Gemini/vLLM 변환)
claude-code-proxy>=0.1.0,<0.2.0

# Web Search MCP
duckduckgo-search>=6.0.0,<8.0.0
beautifulsoup4>=4.12.0,<5.0.0

# Code RAG — tree-sitter AST 파싱 (선택적 — 미설치 시 regex 폴백)
# 정확한 함수/클래스 경계 감지를 위해 설치를 권장합니다.
tree-sitter>=0.23.0,<1.0.0
tree-sitter-python>=0.23.0,<1.0.0
tree-sitter-c>=0.23.0,<1.0.0
tree-sitter-cpp>=0.23.0,<1.0.0
tree-sitter-java>=0.23.0,<1.0.0
